Used the userprofile setting for agreement duration (backwards compatibility/AOM testing). Fixed some invalid unit crashes. Lots of MSVS2005 updates.
git-svn-id: http://ctp2.darkdust.net/anonsvn/trunk@614 34d2fc0c-33df-0310-8a34-85f8040ba873

+++ ctp2_code/gfx/tilesys/tileset.cpp
@@ -69,7 +69,7 @@ extern sint32		g_is565Format;
 
 namespace
 {
-    uint8 const     DIRECTION_INVALID   = -1;
+    uint8 const     DIRECTION_INVALID   = static_cast<uint8>(-1);   // 255
 
     char const      TILESETFILE_555[]   = "gtset555.til";
     char const      TILESETFILE_565[]   = "gtset565.til";
