+++ ctp2_code/ui/aui_directx/aui_directaudiomanager.h
@@ -118,7 +118,7 @@ protected:
 
 	BOOL CreateDSInterface( HWND hwnd );
 	BOOL CreatePrimaryBuffer( void );
-	BOOL SetDSoundRate( ULONG newRate );
+	BOOL SetDSoundRate(DWORD newRate);
 	BOOL InitDS( HWND hwnd );
 };
